refactor chapters & events
